[proliferate] split cloud running shape files
