+++ server/tests/integration/test_cloud_event_sync_api.py
@@ -1,9 +1,5 @@
 from __future__ import annotations
 
-import asyncio
-import json
-from collections.abc import AsyncGenerator, AsyncIterator
-from typing import cast
 from uuid import UUID
 
 import pytest
@@ -15,236 +11,11 @@ from proliferate.db.store.billing import ensure_personal_billing_subject
 from proliferate.db.store.cloud_sync import events as events_store
 from proliferate.db.store.cloud_sync import exposures as exposures_store
 from proliferate.db.store.cloud_sync import projections as projections_store
-from proliferate.integrations.pubsub.models import PubSubMessage
-from proliferate.integrations.pubsub.redis import get_pubsub_bus
-from proliferate.server.cloud.live.domain.channels import target_channel
-from proliferate.server.cloud.live.service import (
-    stream_session_events,
-    stream_target_events,
-    stream_workspace_events,
-)
-from proliferate.server.cloud.workspaces.service import get_cloud_workspace_detail
 from tests.e2e.cloud.helpers.auth import create_user_and_login
-from tests.e2e.cloud.helpers.github import seed_linked_github_account
-from tests.e2e.cloud.helpers.shared import AuthSession
-
-
-async def _create_enrolled_target(
-    client: AsyncClient,
-    db_session: AsyncSession,
-    auth: AuthSession,
-    *,
-    suffix: str = "events",
-) -> tuple[str, dict[str, str]]:
-    await seed_linked_github_account(
-        db_session,
-        user_id=auth.user_id,
-        access_token=f"gh-{suffix}-token",
-    )
-    create = await client.post(
-        "/v1/cloud/targets/enrollments",
-        headers=auth.headers,
-        json={
-            "displayName": f"Event Target {suffix}",
-            "kind": "ssh",
-            "ownerScope": "personal",
-        },
-    )
-    assert create.status_code == 200
-    enrollment = create.json()
-    worker_enroll = await client.post(
-        "/v1/cloud/worker/enroll",
-        json={
-            "enrollmentToken": enrollment["enrollmentToken"],
-            "machineFingerprint": f"{suffix}-machine",
-            "hostname": f"{suffix}-target",
-            "workerVersion": "0.1.0",
-        },
-    )
-    assert worker_enroll.status_code == 200
-    worker = worker_enroll.json()
-    worker_headers = {"Authorization": f"Bearer {worker['workerToken']}"}
-    await _accept_initial_git_identity_command(client, worker_headers)
-    return enrollment["target"]["id"], worker_headers
-
-
-async def _accept_initial_git_identity_command(
-    client: AsyncClient,
-    worker_headers: dict[str, str],
-) -> None:
-    lease = await client.post(
-        "/v1/cloud/worker/commands/lease",
-        headers=worker_headers,
-        json={"supportedKinds": ["configure_git_identity"], "leaseTimeoutSeconds": 300},
-    )
-    assert lease.status_code == 200
-    command = lease.json()["command"]
-    if command is None:
-        return
-    result = await client.post(
-        f"/v1/cloud/worker/commands/{command['commandId']}/result",
-        headers=worker_headers,
-        json={
-            "leaseId": command["leaseId"],
-            "status": "accepted",
-            "result": {"provider": "github"},
-        },
-    )
-    assert result.status_code == 200
-
-
-def _sse_event(frame: str) -> str:
-    for line in frame.splitlines():
-        if line.startswith("event: "):
-            return line.removeprefix("event: ")
-    raise AssertionError(f"SSE frame has no event line: {frame}")
-
-
-def _sse_id(frame: str) -> str:
-    for line in frame.splitlines():
-        if line.startswith("id: "):
-            return line.removeprefix("id: ")
-    raise AssertionError(f"SSE frame has no id line: {frame}")
-
-
-def _sse_data(frame: str) -> dict[str, object]:
-    for line in frame.splitlines():
-        if line.startswith("data: "):
-            value: object = json.loads(line.removeprefix("data: "))
-            assert isinstance(value, dict)
-            return cast("dict[str, object]", value)
-    raise AssertionError(f"SSE frame has no data line: {frame}")
-
-
-async def _next_stream_frame(stream: AsyncIterator[str]) -> str:
-    return await anext(stream)
-
-
-def _mapping(value: object) -> dict[str, object]:
-    assert isinstance(value, dict)
-    return cast("dict[str, object]", value)
-
-
-async def _seed_exposed_session_projection(
-    db_session: AsyncSession,
-    *,
-    target_id: str,
-    auth: AuthSession,
-    workspace_id: str,
-    session_id: str,
-) -> None:
-    target_uuid = UUID(target_id)
-    billing_subject = await ensure_personal_billing_subject(db_session, auth.user_id)
-    workspace = CloudWorkspace(
-        user_id=auth.user_id,
-        owner_scope="personal",
-        owner_user_id=auth.user_id,
-        organization_id=None,
-        created_by_user_id=auth.user_id,
-        billing_subject_id=billing_subject.id,
-        target_id=target_uuid,
-        display_name=f"Workspace {workspace_id}",
-        git_provider="github",
-        git_owner="acme",
-        git_repo_name=workspace_id,
-        normalized_repo_key=f"github/acme/{workspace_id}",
-        git_branch="main",
-        git_base_branch="main",
-        worktree_path=f"/workspace/{workspace_id}",
-        origin="manual_web",
-        origin_json='{"kind":"human","entrypoint":"cloud"}',
-        status="ready",
-        status_detail="Ready",
-        template_version="v1",
-        runtime_generation=0,
-        anyharness_workspace_id=workspace_id,
-        repo_post_ready_phase="idle",
-        repo_post_ready_files_total=0,
-        repo_post_ready_files_applied=0,
-        cleanup_state="none",
-    )
-    db_session.add(workspace)
-    await db_session.flush()
-    exposure = await exposures_store.upsert_workspace_exposure(
-        db_session,
-        target_id=target_uuid,
-        cloud_workspace_id=workspace.id,
-        anyharness_workspace_id=workspace_id,
-        owner_scope="personal",
-        owner_user_id=auth.user_id,
-        organization_id=None,
-        visibility="private",
-        default_projection_level="live",
-        commandable=True,
-        origin="manual_web",
-    )
-    await projections_store.upsert_session_projection_metadata(
-        db_session,
-        target_id=target_uuid,
-        session_id=session_id,
-        exposure_id=exposure.id,
-        cloud_workspace_id=workspace.id,
-        workspace_id=workspace_id,
-        projection_level="live",
-        commandable=True,
-    )
-    await db_session.commit()
-
-
-async def _seed_exposed_workspace(
-    db_session: AsyncSession,
-    *,
-    target_id: str,
-    auth: AuthSession,
-    workspace_id: str,
-) -> tuple[UUID, UUID]:
-    target_uuid = UUID(target_id)
-    billing_subject = await ensure_personal_billing_subject(db_session, auth.user_id)
-    workspace = CloudWorkspace(
-        user_id=auth.user_id,
-        owner_scope="personal",
-        owner_user_id=auth.user_id,
-        organization_id=None,
-        created_by_user_id=auth.user_id,
-        billing_subject_id=billing_subject.id,
-        target_id=target_uuid,
-        display_name=f"Workspace {workspace_id}",
-        git_provider="github",
-        git_owner="acme",
-        git_repo_name=workspace_id,
-        normalized_repo_key=f"github/acme/{workspace_id}",
-        git_branch="main",
-        git_base_branch="main",
-        worktree_path=f"/workspace/{workspace_id}",
-        origin="manual_web",
-        origin_json='{"kind":"human","entrypoint":"cloud"}',
-        status="ready",
-        status_detail="Ready",
-        template_version="v1",
-        runtime_generation=0,
-        anyharness_workspace_id=workspace_id,
-        repo_post_ready_phase="idle",
-        repo_post_ready_files_total=0,
-        repo_post_ready_files_applied=0,
-        cleanup_state="none",
-    )
-    db_session.add(workspace)
-    await db_session.flush()
-    exposure = await exposures_store.upsert_workspace_exposure(
-        db_session,
-        target_id=target_uuid,
-        cloud_workspace_id=workspace.id,
-        anyharness_workspace_id=workspace_id,
-        owner_scope="personal",
-        owner_user_id=auth.user_id,
-        organization_id=None,
-        visibility="private",
-        default_projection_level="live",
-        commandable=True,
-        origin="manual_web",
-    )
-    await db_session.commit()
-    return workspace.id, exposure.id
+from tests.integration.cloud_event_helpers import (
+    create_enrolled_target,
+    seed_exposed_session_projection,
+)
 
 
 class TestCloudEventSyncApi:
@@ -259,8 +30,8 @@ class TestCloudEventSyncApi:
             db_session,
             email_prefix="cloud-event-sync",
         )
-        target_id, worker_headers = await _create_enrolled_target(client, db_session, auth)
-        await _seed_exposed_session_projection(
+        target_id, worker_headers = await create_enrolled_target(client, db_session, auth)
+        await seed_exposed_session_projection(
             db_session,
             target_id=target_id,
             auth=auth,
@@ -432,7 +203,7 @@ class TestCloudEventSyncApi:
             db_session,
             email_prefix="cloud-event-sync-revoked-exposure",
         )
-        target_id, worker_headers = await _create_enrolled_target(
+        target_id, worker_headers = await create_enrolled_target(
             client,
             db_session,
             auth,
@@ -559,14 +330,14 @@ class TestCloudEventSyncApi:
             db_session,
             email_prefix="cloud-event-sync-workspace-mismatch",
         )
-        target_id, worker_headers = await _create_enrolled_target(
+        target_id, worker_headers = await create_enrolled_target(
             client,
             db_session,
             auth,
             suffix="workspace-mismatch",
         )
         target_uuid = UUID(target_id)
-        await _seed_exposed_session_projection(
+        await seed_exposed_session_projection(
             db_session,
             target_id=target_id,
             auth=auth,
@@ -617,307 +388,3 @@ class TestCloudEventSyncApi:
             )
             == 0
         )
-
-    @pytest.mark.asyncio
-    async def test_session_stream_emits_snapshot_and_live_projection_patch(
-        self,
-        client: AsyncClient,
-        db_session: AsyncSession,
-    ) -> None:
-        auth = await create_user_and_login(
-            client,
-            db_session,
-            email_prefix="cloud-event-stream",
-        )
-        target_id, worker_headers = await _create_enrolled_target(
-            client,
-            db_session,
-            auth,
-            suffix="stream",
-        )
-        await _seed_exposed_session_projection(
-            db_session,
-            target_id=target_id,
-            auth=auth,
-            workspace_id="workspace-stream",
-            session_id="session-stream",
-        )
-        uploaded = await client.post(
-            "/v1/cloud/worker/events/batches",
-            headers=worker_headers,
-            json={
-                "events": [
-                    {
-                        "workspaceId": "workspace-stream",
-                        "sessionId": "session-stream",
-                        "seq": 1,
-                        "timestamp": "2026-05-13T00:00:00Z",
-                        "event": {
-                            "type": "session_started",
-                            "nativeSessionId": "native-stream",
-                            "sourceAgentKind": "codex",
-                        },
-                    }
-                ]
-            },
-        )
-        assert uploaded.status_code == 200
-
-        stream = cast(
-            "AsyncGenerator[str, None]",
-            stream_session_events(
-                target_id=UUID(target_id),
-                session_id="session-stream",
-                after_seq=1,
-            ),
-        )
-        try:
-            snapshot_frame = await asyncio.wait_for(anext(stream), timeout=1)
-            assert _sse_event(snapshot_frame) == "snapshot"
-            snapshot = _sse_data(snapshot_frame)
-            assert _mapping(snapshot["session"])["sessionId"] == "session-stream"
-
-            patch_task: asyncio.Task[str] = asyncio.create_task(_next_stream_frame(stream))
-
-            live_uploaded = await client.post(
-                "/v1/cloud/worker/events/batches",
-                headers=worker_headers,
-                json={
-                    "events": [
-                        {
-                            "workspaceId": "workspace-stream",
-                            "sessionId": "session-stream",
-                            "seq": 2,
-                            "timestamp": "2026-05-13T00:00:01Z",
-                            "turnId": "turn-stream",
-                            "itemId": "item-stream",
-                            "event": {
-                                "type": "item_completed",
-                                "item": {
-                                    "kind": "assistant_message",
-                                    "status": "completed",
-                                    "sourceAgentKind": "codex",
-                                    "contentParts": [
-                                        {"type": "text", "text": "streamed response"}
-                                    ],
-                                },
-                            },
-                        }
-                    ]
-                },
-            )
-            assert live_uploaded.status_code == 200
-
-            patch_frame = await asyncio.wait_for(patch_task, timeout=2)
-            assert _sse_event(patch_frame) == "patch"
-            patch_envelope = _sse_data(patch_frame)
-            assert patch_envelope["kind"] == "projection_patch"
-            patch = _mapping(patch_envelope["patch"])
-            assert patch["eventType"] == "item_completed"
-            assert _mapping(patch["transcriptItem"])["text"] == "streamed response"
-        finally:
-            await stream.aclose()
-
-    @pytest.mark.asyncio
-    async def test_workspace_and_target_streams_emit_snapshots_and_patches(
-        self,
-        client: AsyncClient,
-        db_session: AsyncSession,
-    ) -> None:
-        auth = await create_user_and_login(
-            client,
-            db_session,
-            email_prefix="cloud-workspace-stream",
-        )
-        target_id, worker_headers = await _create_enrolled_target(
-            client,
-            db_session,
-            auth,
-            suffix="workspace-stream",
-        )
-        cloud_workspace_uuid, exposure_id = await _seed_exposed_workspace(
-            db_session,
-            target_id=target_id,
-            auth=auth,
-            workspace_id="workspace-live",
-        )
-        cloud_workspace_id = str(cloud_workspace_uuid)
-        target_stream = cast(
-            "AsyncGenerator[str, None]",
-            stream_target_events(target_id=UUID(target_id), after_seq=0),
-        )
-        try:
-            target_snapshot_frame = await asyncio.wait_for(anext(target_stream), timeout=1)
-            assert _sse_event(target_snapshot_frame) == "snapshot"
-            assert _mapping(_sse_data(target_snapshot_frame)["target"])["id"] == target_id
-
-            cursor_task: asyncio.Task[str] = asyncio.create_task(_next_stream_frame(target_stream))
-            bus = get_pubsub_bus()
-            await bus.publish(
-                target_channel(target_id=UUID(target_id)),
-                PubSubMessage(event="patch", event_id="0", data={"kind": "stale_patch"}),
-            )
-            await asyncio.sleep(0.05)
-            assert not cursor_task.done()
-            await bus.publish(
-                target_channel(target_id=UUID(target_id)),
-                PubSubMessage(event="patch", event_id="1", data={"kind": "fresh_patch"}),
-            )
-            cursor_frame = await asyncio.wait_for(cursor_task, timeout=1)
-            assert _sse_event(cursor_frame) == "patch"
-            assert _sse_id(cursor_frame) == "1"
-            assert _sse_data(cursor_frame)["kind"] == "fresh_patch"
-
-            target_patch_task: asyncio.Task[str] = asyncio.create_task(
-                _next_stream_frame(target_stream)
-            )
-            heartbeat = await client.post(
-                "/v1/cloud/worker/heartbeat",
-                headers=worker_headers,
-                json={"status": "online", "statusDetail": "ready"},
-            )
-            assert heartbeat.status_code == 200
-            target_patch_frame = await asyncio.wait_for(target_patch_task, timeout=2)
-            assert _sse_event(target_patch_frame) == "patch"
-            target_patch = _sse_data(target_patch_frame)
-            assert target_patch["kind"] == "target_projection_patch"
-            assert _mapping(target_patch["target"])["status"] == "online"
-
-            command_status_task: asyncio.Task[str] = asyncio.create_task(
-                _next_stream_frame(target_stream)
-            )
-            command = await client.post(
-                "/v1/cloud/commands",
-                headers=auth.headers,
-                json={
-                    "idempotencyKey": "target-stream-command",
-                    "targetId": target_id,
-                    "workspaceId": "workspace-live",
-                    "cloudWorkspaceId": cloud_workspace_id,
-                    "kind": "backfill_exposed_workspace",
-                    "payload": {},
-                    "source": "desktop_cloud_view",
-                },
-            )
-            assert command.status_code == 200
-            command_status_frame = await asyncio.wait_for(command_status_task, timeout=2)
-            assert _sse_event(command_status_frame) == "command_status"
-            command_status = _sse_data(command_status_frame)
-            assert command_status["kind"] == "command_status"
-            assert _mapping(command_status["command"])["status"] == "queued"
-        finally:
-            await target_stream.aclose()
-
-        backfill = await client.post(
-            "/v1/cloud/worker/backfill",
-            headers=worker_headers,
-            json={
-                "workspaces": [
-                    {
-                        "workspaceId": "workspace-live",
-                        "displayName": "Live Workspace",
-                        "repo": {
-                            "provider": "github",
-                            "owner": "proliferate-ai",
-                            "name": "proliferate",
-                            "branch": "main",
-                            "baseBranch": "main",
-                        },
-                    }
-                ],
-                "sessions": [
-                    {
-                        "workspaceId": "workspace-live",
-                        "sessionId": "session-live",
-                        "sourceAgentKind": "codex",
-                        "status": "running",
-                    }
-                ],
-            },
-        )
-        assert backfill.status_code == 200
-        assert backfill.json()["mappedWorkspaces"][0]["cloudWorkspaceId"] == cloud_workspace_id
-        await projections_store.upsert_session_projection_metadata(
-            db_session,
-            target_id=UUID(target_id),
-            session_id="session-live",
-            exposure_id=exposure_id,
-            cloud_workspace_id=UUID(cloud_workspace_id),
-            workspace_id="workspace-live",
-            projection_level="live",
-            commandable=True,
-        )
-        await db_session.commit()
-
-        uploaded = await client.post(
-            "/v1/cloud/worker/events/batches",
-            headers=worker_headers,
-            json={
-                "events": [
-                    {
-                        "workspaceId": "workspace-live",
-                        "sessionId": "session-live",
-                        "seq": 1,
-                        "timestamp": "2026-05-13T00:00:00Z",
-                        "event": {
-                            "type": "session_started",
-                            "sourceAgentKind": "codex",
-                        },
-                    }
-                ]
-            },
-        )
-        assert uploaded.status_code == 200
-
-        workspace_snapshot_response = await client.get(
-            f"/v1/cloud/workspaces/{cloud_workspace_id}/snapshot",
-            headers=auth.headers,
-        )
-        assert workspace_snapshot_response.status_code == 200
-        workspace_snapshot = workspace_snapshot_response.json()
-        assert workspace_snapshot["workspace"]["id"] == cloud_workspace_id
-        assert workspace_snapshot["sessions"][0]["sessionId"] == "session-live"
-
-        workspace = await get_cloud_workspace_detail(
-            db_session,
-            UUID(auth.user_id),
-            UUID(cloud_workspace_id),
-        )
-        workspace_stream = cast(
-            "AsyncGenerator[str, None]",
-            stream_workspace_events(workspace=workspace, after_seq=1),
-        )
-        try:
-            workspace_snapshot_frame = await asyncio.wait_for(anext(workspace_stream), timeout=1)
-            assert _sse_event(workspace_snapshot_frame) == "snapshot"
-            workspace_snapshot = _sse_data(workspace_snapshot_frame)
-            assert _mapping(workspace_snapshot["workspace"])["id"] == cloud_workspace_id
-            assert _mapping(workspace_snapshot["sessions"][0])["sessionId"] == "session-live"
-
-            workspace_patch_task: asyncio.Task[str] = asyncio.create_task(
-                _next_stream_frame(workspace_stream)
-            )
-            live_uploaded = await client.post(
-                "/v1/cloud/worker/events/batches",
-                headers=worker_headers,
-                json={
-                    "events": [
-                        {
-                            "workspaceId": "workspace-live",
-                            "sessionId": "session-live",
-                            "seq": 2,
-                            "timestamp": "2026-05-13T00:00:01Z",
-                            "event": {"type": "turn_started"},
-                        }
-                    ]
-                },
-            )
-            assert live_uploaded.status_code == 200
-
-            workspace_patch_frame = await asyncio.wait_for(workspace_patch_task, timeout=2)
-            assert _sse_event(workspace_patch_frame) == "patch"
-            patch_envelope = _sse_data(workspace_patch_frame)
-            assert patch_envelope["kind"] == "workspace_projection_patch"
-            assert _mapping(patch_envelope["patch"])["eventType"] == "turn_started"
-        finally:
-            await workspace_stream.aclose()
